+++ packages/spacecat-shared-data-access/test/unit/models/audit/audit.model.test.js
@@ -241,6 +241,7 @@ describe('AuditModel', () => {
       expect(formattedPayload).to.deep.equal({
         type: 'someType',
         siteId: 'someSiteId',
+        allowCache: true,
         auditContext: { some: 'context' },
       });
     });
@@ -251,14 +252,22 @@ describe('AuditModel', () => {
         siteId: 'someSiteId',
         processingType: 'someProcessingType',
       };
+      const context = {
+        env: {
+          AUDIT_JOBS_QUEUE_URL: 'audit-jobs-queue-url',
+        },
+      };
       const auditContext = { some: 'context' };
       const formattedPayload = auditStepDestinationConfigs[auditStepDestinations.CONTENT_SCRAPER]
-        .formatPayload(stepResult, auditContext);
+        .formatPayload(stepResult, auditContext, context);
 
       expect(formattedPayload).to.deep.equal({
         urls: [{ url: 'someUrl' }],
         jobId: 'someSiteId',
         processingType: 'someProcessingType',
+        completionQueueUrl: 'audit-jobs-queue-url',
+        skipMessage: false,
+        allowCache: true,
         auditContext: { some: 'context' },
       });
     });