+++ packages/spacecat-shared-example/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-example-v1.2.24](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-example-v1.2.23...@adobe/spacecat-shared-example-v1.2.24) (2025-04-15)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-universal to v5.2.0 ([#694](https://github.com/adobe/spacecat-shared/issues/694)) ([55883c5](https://github.com/adobe/spacecat-shared/commit/55883c597c61c891fc17ed39d0aab1c33af5b90c))
+
 # [@adobe/spacecat-shared-example-v1.2.23](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-example-v1.2.22...@adobe/spacecat-shared-example-v1.2.23) (2025-04-01)