+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,31 @@
+# [@adobe/spacecat-shared-google-client-v1.4.21](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.20...@adobe/spacecat-shared-google-client-v1.4.21) (2025-05-03)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#712](https://github.com/adobe/spacecat-shared/issues/712)) ([76cebd2](https://github.com/adobe/spacecat-shared/commit/76cebd2a7a7b9799e4ca265833620eada01f5c8c))
+
+# [@adobe/spacecat-shared-google-client-v1.4.20](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.19...@adobe/spacecat-shared-google-client-v1.4.20) (2025-04-26)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#708](https://github.com/adobe/spacecat-shared/issues/708)) ([57535ac](https://github.com/adobe/spacecat-shared/commit/57535ac8c636de229cec7c5cee83dead07ac09fb))
+
+# [@adobe/spacecat-shared-google-client-v1.4.19](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.18...@adobe/spacecat-shared-google-client-v1.4.19) (2025-04-15)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-universal to v5.2.0 ([#694](https://github.com/adobe/spacecat-shared/issues/694)) ([55883c5](https://github.com/adobe/spacecat-shared/commit/55883c597c61c891fc17ed39d0aab1c33af5b90c))
+
+# [@adobe/spacecat-shared-google-client-v1.4.18](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.17...@adobe/spacecat-shared-google-client-v1.4.18) (2025-04-15)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#679](https://github.com/adobe/spacecat-shared/issues/679)) ([a41bf0c](https://github.com/adobe/spacecat-shared/commit/a41bf0cd488efa0f72af0933992edb256302af18))
+
 # [@adobe/spacecat-shared-google-client-v1.4.17](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.16...@adobe/spacecat-shared-google-client-v1.4.17) (2025-04-01)